+++ coupled_simulation/__init__.py
@@ -1,0 +1,10 @@
+"""
+.. moduleauthor:: Francesco Witte <francesco.witte@hs-flensburg.de>
+
+"""
+__version__ = '0.0.1 dev'
+
+from coupled_simulation import coupling as cp
+from coupled_simulation import powerplant as pp
+from coupled_simulation import geostorage as gs
+from coupled_simulation import utilities as util
