--- coupling/__init__.py
@@ -1,9 +1,0 @@
-"""
-.. moduleauthor:: Francesco Witte <francesco.witte@hs-flensburg.de>
-
-"""
-__version__ = '0.0.1 dev'
-
-from coupling import coupling as cp
-from coupling import powerplant as pp
-from coupling import geostorage as gs
