+++ nose.cfg
@@ -1,3 +1,3 @@
 [nosetests]
 with-coverage=1
-cover-package=coupling
+cover-package=coupled_simulation
